+++ Plan/common/src/main/java/com/djrapitops/plan/PlanSystem.java
@@ -293,6 +293,13 @@ public class PlanSystem implements SubSystem {
         return extensionService;
     }
 
+    /**
+     * Originally visible for testing purposes.
+     *
+     * @return the error logger of the system
+     * @deprecated A smell, dagger should be used to construct things instead.
+     */
+    @Deprecated
     public ErrorLogger getErrorLogger() {
         return errorLogger;
     }
