+++ Plan/common/src/main/java/com/djrapitops/plan/utilities/logging/ErrorContext.java
@@ -81,4 +81,12 @@ public class ErrorContext implements Serializable {
             return context;
         }
     }
+
+    @Override
+    public String toString() {
+        return "ErrorContext{" +
+                "related=" + related +
+                ", whatToDo='" + whatToDo + '\'' +
+                '}';
+    }
 }
