+++ Plan/common/src/test/java/com/djrapitops/plan/storage/database/DatabaseTestComponent.java
@@ -14,51 +14,64 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with Plan. If not, see <https://www.gnu.org/licenses/>.
  */
-package utilities.dagger;
+package com.djrapitops.plan.storage.database;
 
-import com.djrapitops.plan.PlanPlugin;
-import com.djrapitops.plan.PlanSystem;
-import com.djrapitops.plan.commands.PlanCommand;
-import com.djrapitops.plan.modules.PlaceholderModule;
-import com.djrapitops.plan.utilities.logging.PluginErrorLogger;
+import com.djrapitops.plan.delivery.DeliveryUtilities;
+import com.djrapitops.plan.identification.ServerInfo;
+import com.djrapitops.plan.settings.ConfigSystem;
+import com.djrapitops.plan.settings.config.PlanConfig;
+import com.djrapitops.plan.storage.file.PlanFiles;
 import dagger.BindsInstance;
 import dagger.Component;
+import utilities.DBPreparer;
+import utilities.dagger.*;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
 import java.nio.file.Path;
 
-/**
- * Dagger component for {@link com.djrapitops.plan.PlanPlugin} based Plan system.
- *
- * @author Rsl1122
- */
 @Singleton
 @Component(modules = {
-        PlanPluginModule.class,
+        DBSystemModule.class,
         TestSystemObjectProvidingModule.class,
+
         TestAPFModule.class,
-        PlaceholderModule.class,
-
+        PlanPluginModule.class,
         PluginServerPropertiesModule.class,
-        PluginSuperClassBindingModule.class,
-        DBSystemModule.class
+        PluginSuperClassBindingModule.class
 })
-public interface PlanPluginComponent {
-    PlanCommand planCommand();
+public interface DatabaseTestComponent extends DBPreparer.Dependencies {
 
-    PlanSystem system();
+    default void enable() {
+        configSystem().enable();
+        dbSystem().enable();
+        serverInfo().enable();
+    }
 
-    PluginErrorLogger pluginErrorLogger();
+    default void disable() {
+        serverInfo().disable();
+        dbSystem().disable();
+        configSystem().disable();
+    }
+
+    PlanConfig config();
+
+    ConfigSystem configSystem();
+
+    DBSystem dbSystem();
+
+    ServerInfo serverInfo();
+
+    DeliveryUtilities deliveryUtilities();
+
+    PlanFiles files();
 
     @Component.Builder
     interface Builder {
         @BindsInstance
         Builder bindTemporaryDirectory(@Named("tempDir") Path tempDir);
 
-        @BindsInstance
-        Builder plan(PlanPlugin plan);
-
-        PlanPluginComponent build();
+        DatabaseTestComponent build();
     }
+
 }
