+++ Plan/common/src/test/java/com/djrapitops/plan/storage/database/DatabaseTestPreparer.java
@@ -17,6 +17,9 @@
 package com.djrapitops.plan.storage.database;
 
 import com.djrapitops.plan.PlanSystem;
+import com.djrapitops.plan.delivery.DeliveryUtilities;
+import com.djrapitops.plan.identification.ServerInfo;
+import com.djrapitops.plan.settings.config.PlanConfig;
 import com.djrapitops.plan.storage.database.transactions.Executable;
 import com.djrapitops.plan.storage.database.transactions.Transaction;
 import utilities.TestConstants;
@@ -33,7 +36,24 @@ public interface DatabaseTestPreparer {
 
     UUID serverUUID();
 
+    @Deprecated
     PlanSystem system();
+
+    default PlanConfig config() {
+        return system().getConfigSystem().getConfig();
+    }
+
+    default DBSystem dbSystem() {
+        return system().getDatabaseSystem();
+    }
+
+    default ServerInfo serverInfo() {
+        return system().getServerInfo();
+    }
+
+    default DeliveryUtilities deliveryUtilities() {
+        return system().getDeliveryUtilities();
+    }
 
     default void execute(Executable executable) {
         db().executeTransaction(new Transaction() {
