+++ Plan/common/src/test/java/com/djrapitops/plan/storage/database/H2Test.java
@@ -17,7 +17,10 @@
 package com.djrapitops.plan.storage.database;
 
 import com.djrapitops.plan.PlanSystem;
+import com.djrapitops.plan.delivery.DeliveryUtilities;
 import com.djrapitops.plan.identification.Server;
+import com.djrapitops.plan.identification.ServerInfo;
+import com.djrapitops.plan.settings.config.PlanConfig;
 import com.djrapitops.plan.storage.database.queries.*;
 import com.djrapitops.plan.storage.database.transactions.StoreServerInformationTransaction;
 import com.djrapitops.plan.storage.database.transactions.commands.RemoveEverythingTransaction;
@@ -28,15 +31,16 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import utilities.DBPreparer;
 import utilities.RandomData;
-import utilities.mocks.PluginMockComponent;
 
 import java.nio.file.Path;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
 
 /**
  * Tests for the H2 database.
@@ -61,13 +65,16 @@ public class H2Test implements DatabaseTest,
 
     private static final int TEST_PORT_NUMBER = RandomData.randomInt(9005, 9500);
 
-    private static PlanSystem system;
     private static Database database;
+    private static DatabaseTestComponent component;
 
     @BeforeAll
     static void setupDatabase(@TempDir Path temp) throws Exception {
-        system = new PluginMockComponent(temp).getPlanSystem();
-        database = new DBPreparer(system, TEST_PORT_NUMBER).prepareH2()
+        component = DaggerDatabaseTestComponent.builder()
+                .bindTemporaryDirectory(temp)
+                .build();
+        component.enable();
+        database = new DBPreparer(component, TEST_PORT_NUMBER).prepareH2()
                 .orElseThrow(IllegalStateException::new);
     }
 
@@ -101,7 +108,7 @@ public class H2Test implements DatabaseTest,
             database.close();
             System.out.println("Database state after close: " + database.getState().name());
         }
-        system.disable();
+        component.disable();
     }
 
     @Override
@@ -111,11 +118,33 @@ public class H2Test implements DatabaseTest,
 
     @Override
     public UUID serverUUID() {
-        return system.getServerInfo().getServerUUID();
+        return component.serverInfo().getServerUUID();
+    }
+
+    @Override
+    public PlanConfig config() {
+        return component.config();
+    }
+
+    @Override
+    public DBSystem dbSystem() {
+        return component.dbSystem();
+    }
+
+    @Override
+    public ServerInfo serverInfo() {
+        return component.serverInfo();
+    }
+
+    @Override
+    public DeliveryUtilities deliveryUtilities() {
+        return component.deliveryUtilities();
     }
 
     @Override
     public PlanSystem system() {
-        return system;
+        PlanSystem mockSystem = Mockito.mock(PlanSystem.class);
+        when(mockSystem.getPlanFiles()).thenReturn(component.files());
+        return mockSystem;
     }
 }
