+++ Plan/common/src/test/java/com/djrapitops/plan/storage/database/MySQLTest.java
@@ -17,7 +17,10 @@
 package com.djrapitops.plan.storage.database;
 
 import com.djrapitops.plan.PlanSystem;
+import com.djrapitops.plan.delivery.DeliveryUtilities;
 import com.djrapitops.plan.identification.Server;
+import com.djrapitops.plan.identification.ServerInfo;
+import com.djrapitops.plan.settings.config.PlanConfig;
 import com.djrapitops.plan.storage.database.queries.*;
 import com.djrapitops.plan.storage.database.transactions.StoreServerInformationTransaction;
 import com.djrapitops.plan.storage.database.transactions.commands.RemoveEverythingTransaction;
@@ -29,16 +32,17 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import utilities.DBPreparer;
 import utilities.RandomData;
-import utilities.mocks.PluginMockComponent;
 
 import java.nio.file.Path;
 import java.util.Optional;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
 
 /**
  * Tests for MySQL database.
@@ -67,13 +71,16 @@ class MySQLTest implements DatabaseTest,
 
     private static final int TEST_PORT_NUMBER = RandomData.randomInt(9005, 9500);
 
-    private static PlanSystem system;
     private static Database database;
+    private static DatabaseTestComponent component;
 
     @BeforeAll
     static void setupDatabase(@TempDir Path temp) throws Exception {
-        system = new PluginMockComponent(temp).getPlanSystem();
-        Optional<Database> mysql = new DBPreparer(system, TEST_PORT_NUMBER).prepareMySQL();
+        component = DaggerDatabaseTestComponent.builder()
+                .bindTemporaryDirectory(temp)
+                .build();
+        component.enable();
+        Optional<Database> mysql = new DBPreparer(component, TEST_PORT_NUMBER).prepareMySQL();
         Assumptions.assumeTrue(mysql.isPresent());
         database = mysql.get();
     }
@@ -104,7 +111,7 @@ class MySQLTest implements DatabaseTest,
     @AfterAll
     static void disableSystem() {
         if (database != null) database.close();
-        system.disable();
+        component.disable();
     }
 
     @Override
@@ -114,11 +121,33 @@ class MySQLTest implements DatabaseTest,
 
     @Override
     public UUID serverUUID() {
-        return system.getServerInfo().getServerUUID();
+        return component.serverInfo().getServerUUID();
+    }
+
+    @Override
+    public PlanConfig config() {
+        return component.config();
+    }
+
+    @Override
+    public DBSystem dbSystem() {
+        return component.dbSystem();
+    }
+
+    @Override
+    public ServerInfo serverInfo() {
+        return component.serverInfo();
+    }
+
+    @Override
+    public DeliveryUtilities deliveryUtilities() {
+        return component.deliveryUtilities();
     }
 
     @Override
     public PlanSystem system() {
-        return system;
+        PlanSystem mockSystem = Mockito.mock(PlanSystem.class);
+        when(mockSystem.getPlanFiles()).thenReturn(component.files());
+        return mockSystem;
     }
 }
